insert macro modification in KomaUI.jl
