+++ src/KomaUI.jl
@@ -5,6 +5,24 @@ obj_ui = Observable{Phantom{Float64}}(Phantom{Float64}(x=[0.0]))
 raw_ui = Observable{RawAcquisitionData}(setup_raw())
 img_ui = Observable{Array{ComplexF64}}([0.0im 0.; 0. 0.])
 
+macro unsafe_blink()
+    Core.eval(@__MODULE__, quote 
+        function Blink.AtomShell.init(; debug = false)
+            Blink.AtomShell.electron() # Check path exists
+            p, dp = Blink.AtomShell.port(), Blink.AtomShell.port()
+            debug && Blink.AtomShell.inspector(dp)
+            dbg = debug ? "--debug=$dp" : []
+            proc = (debug ? Blink.AtomShell.run_rdr : Blink.AtomShell.run)(
+                `$(Blink.AtomShell.electron()) --no-sandbox $dbg $(Blink.AtomShell.mainjs) port $p`; wait=false)
+            conn = Blink.AtomShell.try_connect(ip"127.0.0.1", p)
+            shell = Blink.AtomShell.Electron(proc, conn)
+            Blink.AtomShell.initcbs(shell)
+            return shell
+        end
+    end)
+    return nothing  # Avoid returning an unevaluated expression
+end
+
 """
     out = KomaUI(; kwargs...)
